+++ ietfparse/algorithms.py
@@ -35,6 +35,8 @@ def _content_type_matches(
 def select_content_type(  # noqa: C901 -- overly complex
     requested: abc.Sequence[datastructures.ContentType | str] | str,
     available: abc.Sequence[datastructures.ContentType | str],
+    *,
+    default: datastructures.ContentType | str | None = None,
 ) -> tuple[datastructures.ContentType, datastructures.ContentType]:
     """Select the best content type.
 
@@ -53,10 +55,14 @@ def select_content_type(  # noqa: C901 -- overly complex
     :param available: a sequence of
         [ietfparse.datastructures.ContentType][] instances that the
         server is capable of producing
+    :param default: optional default value to return if there is
+        no acceptable match
     :returns: the selected content type (from `available`) and the
         pattern that it matched (from `requested`)
 
-    raises [ietfparse.errors.NoMatch][] when a suitable match was not found
+    :raises ietfparse.errors.NoMatch: when a suitable match was not found
+    :raises ValueError: when `default` is specified and it is not in
+        `available`
 
     """
 
@@ -119,21 +125,8 @@ def select_content_type(  # noqa: C901 -- overly complex
     def extract_quality(obj: datastructures.ContentType) -> float:
         return 1.0 if obj.quality is None else obj.quality
 
-    _requested: abc.Sequence[datastructures.ContentType]
-    if isinstance(requested, str):
-        _requested = _helpers.parse_header('parse_accept', requested)
-    else:
-        _requested = [
-            value
-            if isinstance(value, datastructures.ContentType)
-            else _helpers.parse_header('parse_content_type', value)
-            for value in requested
-        ]
-    _available = sorted(
-        value
-        if isinstance(value, datastructures.ContentType)
-        else _helpers.parse_header('parse_content_type', value)
-        for value in available
+    _requested, _available, _default = _normalize_parameters(
+        requested, available, default
     )
 
     matches = []
@@ -147,9 +140,45 @@ def select_content_type(  # noqa: C901 -- overly complex
                 matches.append(Match(candidate, pattern))
 
     if not matches:
+        if _default is not None:
+            return _default, _default
         raise errors.NoMatch
 
     matches = sorted(
         matches, key=attrgetter('match_type', 'parameter_distance')
     )
     return matches[0].candidate, matches[0].pattern
+
+
+def _normalize_parameters(
+    requested: abc.Sequence[datastructures.ContentType | str] | str,
+    available: abc.Sequence[datastructures.ContentType | str],
+    default: datastructures.ContentType | str | None,
+) -> tuple[
+    abc.Sequence[datastructures.ContentType],
+    abc.Sequence[datastructures.ContentType],
+    datastructures.ContentType | None,
+]:
+    if isinstance(requested, str):
+        _requested = _helpers.parse_header('parse_accept', requested)
+    else:
+        _requested = [
+            r
+            if isinstance(r, datastructures.ContentType)
+            else _helpers.parse_header('parse_content_type', r)
+            for r in requested
+        ]
+
+    _available = [
+        a
+        if isinstance(a, datastructures.ContentType)
+        else _helpers.parse_header('parse_content_type', a)
+        for a in available
+    ]
+
+    if isinstance(default, str):
+        default = _helpers.parse_header('parse_content_type', default)
+        if default not in _available:
+            raise ValueError('default content type not in available')
+
+    return _requested, sorted(_available), default